+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 - Black background
 # v2
 - Add a single, stationary celestial body
+# v3
+- Let celestial body position, color be configured